Move projets-validator in /server
